+++ src/plugins/map_generation/src/traits.rs
@@ -27,7 +27,7 @@ pub(crate) trait ExtraComponentsDefinition {
 }
 
 pub(crate) trait GridCellDistanceDefinition {
-	const CELL_DISTANCE: u8;
+	const CELL_DISTANCE: f32;
 }
 
 pub(crate) trait Spawn {
